Add per-partition job info and absolute slurm paths
Per-partition job status totals are recorded: running, configuring,
completing, pending.

When run via cron the script could fail as PATH isn't guaranteed
to include the directory containing the slurm commands.  There is
now a variable that can be set at the top of the script.  Commands
are then run as absolute path commands.